Factor cube rendering into a hook
