+++ src/modules/hooks/useCube.ts
@@ -1,0 +1,305 @@
+import React, { MutableRefObject } from 'react';
+import { mat4 } from 'gl-matrix';
+
+import useAnimationFrame from './useAnimationFrame';
+
+interface ProgramInfo {
+    program: WebGLProgram,
+    attribLocations: {
+        vertexPosition: number,
+        vertexColor: number,
+    },
+    uniformLocations: {
+        projectionMatrix: WebGLUniformLocation,
+        modelViewMatrix: WebGLUniformLocation,
+    },
+}
+
+const positions = [
+    // Front face
+    -1.0, -1.0,  1.0,
+     1.0, -1.0,  1.0,
+     1.0,  1.0,  1.0,
+    -1.0,  1.0,  1.0,
+    
+    // Back face
+    -1.0, -1.0, -1.0,
+    -1.0,  1.0, -1.0,
+     1.0,  1.0, -1.0,
+     1.0, -1.0, -1.0,
+    
+    // Top face
+    -1.0,  1.0, -1.0,
+    -1.0,  1.0,  1.0,
+     1.0,  1.0,  1.0,
+     1.0,  1.0, -1.0,
+    
+    // Bottom face
+    -1.0, -1.0, -1.0,
+     1.0, -1.0, -1.0,
+     1.0, -1.0,  1.0,
+    -1.0, -1.0,  1.0,
+    
+    // Right face
+     1.0, -1.0, -1.0,
+     1.0,  1.0, -1.0,
+     1.0,  1.0,  1.0,
+     1.0, -1.0,  1.0,
+    
+    // Left face
+    -1.0, -1.0, -1.0,
+    -1.0, -1.0,  1.0,
+    -1.0,  1.0,  1.0,
+    -1.0,  1.0, -1.0,
+];
+
+const faceColors = [
+    [1.0,  1.0,  1.0,  1.0],    // Front face: white
+    [1.0,  0.0,  0.0,  1.0],    // Back face: red
+    [0.0,  1.0,  0.0,  1.0],    // Top face: green
+    [0.0,  0.0,  1.0,  1.0],    // Bottom face: blue
+    [1.0,  1.0,  0.0,  1.0],    // Right face: yellow
+    [1.0,  0.0,  1.0,  1.0],    // Left face: purple
+];
+
+const indices = [
+    0,  1,  2,      0,  2,  3,    // front
+    4,  5,  6,      4,  6,  7,    // back
+    8,  9,  10,     8,  10, 11,   // top
+    12, 13, 14,     12, 14, 15,   // bottom
+    16, 17, 18,     16, 18, 19,   // right
+    20, 21, 22,     20, 22, 23,   // left
+];
+
+const vsSource = `
+    attribute vec4 aVertexPosition;
+    attribute vec4 aVertexColor;
+    uniform mat4 uModelViewMatrix;
+    uniform mat4 uProjectionMatrix;
+    varying lowp vec4 vColor;
+    void main(void) {
+      gl_Position = uProjectionMatrix * uModelViewMatrix * aVertexPosition;
+      vColor = aVertexColor;
+    }
+`;
+
+const fsSource = `
+    varying lowp vec4 vColor;
+    void main(void) {
+      gl_FragColor = vColor;
+    }
+`;
+
+function loadShader(gl: WebGLRenderingContext, type: number, source: string) {
+    const shader = gl.createShader(type);
+  
+    // Send the source to the shader object
+  
+    gl.shaderSource(shader, source);
+  
+    // Compile the shader program
+  
+    gl.compileShader(shader);
+  
+    // See if it compiled successfully
+  
+    if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
+      alert('An error occurred compiling the shaders: ' + gl.getShaderInfoLog(shader));
+      gl.deleteShader(shader);
+      return null;
+    }
+  
+    return shader;
+}
+
+function initShaderProgram(gl: WebGLRenderingContext, vsSource: string, fsSource: string) {
+    const vertexShader = loadShader(gl, gl.VERTEX_SHADER, vsSource);
+    const fragmentShader = loadShader(gl, gl.FRAGMENT_SHADER, fsSource);
+  
+    // Create the shader program
+  
+    const shaderProgram = gl.createProgram();
+    gl.attachShader(shaderProgram, vertexShader);
+    gl.attachShader(shaderProgram, fragmentShader);
+    gl.linkProgram(shaderProgram);
+  
+    // If creating the shader program failed, alert
+  
+    if (!gl.getProgramParameter(shaderProgram, gl.LINK_STATUS)) {
+      alert('Unable to initialize the shader program: ' + gl.getProgramInfoLog(shaderProgram));
+      return null;
+    }
+  
+    return shaderProgram;
+}
+
+type Hook = (contextRef: MutableRefObject<WebGLRenderingContext>) => void;
+
+const useCube: Hook = contextRef => {
+    const programInfoRef = React.useRef<ProgramInfo>(null);
+    const positionBufferRef = React.useRef<WebGLBuffer>(null);
+    const colorBufferRef = React.useRef<WebGLBuffer>(null);
+    const indexBufferRef = React.useRef<WebGLBuffer>(null);
+    const rotationRef = React.useRef<number>(0);
+
+    function draw(delta: number) {
+        const gl = contextRef.current;
+        const programInfo = programInfoRef.current;
+        const positionBuffer = positionBufferRef.current;
+        const colorBuffer = colorBufferRef.current;
+        const indexBuffer = indexBufferRef.current;
+
+        if (!gl || !programInfo || !positionBuffer || !colorBuffer || !indexBuffer) {
+            return;
+        }
+
+        // Delta in seconds.
+        rotationRef.current += delta * 0.001;
+
+        gl.viewport(0, 0, gl.canvas.width, gl.canvas.height);
+        gl.clearColor(0.9, 0.9, 0.9, 1.0);
+        gl.clearDepth(1.0);
+        gl.enable(gl.DEPTH_TEST);
+        gl.depthFunc(gl.LEQUAL);
+
+        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+
+        const fieldOfView = 45 * Math.PI / 180;   // in radians
+        const aspect = gl.canvas.width / gl.canvas.height;
+        const zNear = 0.1;
+        const zFar = 100.0;
+        const projectionMatrix = mat4.create();   
+
+        mat4.perspective(
+            projectionMatrix,
+            fieldOfView,
+            aspect,
+            zNear,
+            zFar);
+
+        const modelViewMatrix = mat4.create();
+
+        mat4.translate(modelViewMatrix,     // destination matrix
+            modelViewMatrix,     // matrix to translate
+            [-0.0, 0.0, -6.0]);  // amount to translate
+        mat4.rotate(modelViewMatrix,  // destination matrix
+            modelViewMatrix,  // matrix to rotate
+            rotationRef.current,     // amount to rotate in radians
+            [0, 0, 1]);       // axis to rotate around (Z)
+        mat4.rotate(modelViewMatrix,  // destination matrix
+            modelViewMatrix,  // matrix to rotate
+            rotationRef.current * .7,// amount to rotate in radians
+            [0, 1, 0]);       // axis to rotate around (X)
+
+        {
+            const numComponents = 3;
+            const type = gl.FLOAT;
+            const normalize = false;
+            const stride = 0;
+            const offset = 0;
+    
+            gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
+            gl.vertexAttribPointer(
+                programInfo.attribLocations.vertexPosition,
+                numComponents,
+                type,
+                normalize,
+                stride,
+                offset);
+            gl.enableVertexAttribArray(
+                programInfo.attribLocations.vertexPosition);
+        }
+
+        {
+            const numComponents = 4;
+            const type = gl.FLOAT;
+            const normalize = false;
+            const stride = 0;
+            const offset = 0;
+            gl.bindBuffer(gl.ARRAY_BUFFER, colorBuffer);
+            gl.vertexAttribPointer(
+                programInfo.attribLocations.vertexColor,
+                numComponents,
+                type,
+                normalize,
+                stride,
+                offset);
+            gl.enableVertexAttribArray(
+                programInfo.attribLocations.vertexColor);
+        }
+
+        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, indexBuffer);
+
+        gl.useProgram(programInfo.program);
+
+        gl.uniformMatrix4fv(
+            programInfo.uniformLocations.projectionMatrix,
+            false,
+            projectionMatrix);
+        gl.uniformMatrix4fv(
+            programInfo.uniformLocations.modelViewMatrix,
+            false,
+            modelViewMatrix);
+        
+        {
+            const vertexCount = 36;
+            const type = gl.UNSIGNED_SHORT;
+            const offset = 0;
+            gl.drawElements(gl.TRIANGLES, vertexCount, type, offset);
+        }
+    }
+
+    function createCube() {
+        const gl = contextRef.current;
+        if (!gl) {
+            return;
+        }
+
+        const shaderProgram = initShaderProgram(gl, vsSource, fsSource);
+
+        const programInfo = {
+            program: shaderProgram,
+            attribLocations: {
+              vertexPosition: gl.getAttribLocation(shaderProgram, 'aVertexPosition'),
+              vertexColor: gl.getAttribLocation(shaderProgram, 'aVertexColor'),
+            },
+            uniformLocations: {
+              projectionMatrix: gl.getUniformLocation(shaderProgram, 'uProjectionMatrix'),
+              modelViewMatrix: gl.getUniformLocation(shaderProgram, 'uModelViewMatrix'),
+            },
+        };
+
+        programInfoRef.current = programInfo;
+
+        const positionBuffer = gl.createBuffer();
+        positionBufferRef.current = positionBuffer;
+        gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(positions), gl.STATIC_DRAW);
+        
+        let colors: number[] = [];
+        for (var j = 0; j < faceColors.length; ++j) {
+            const c = faceColors[j];
+        
+            // Repeat each color four times for the four vertices of the face
+            colors = colors.concat(c, c, c, c);
+        }
+
+        const colorBuffer = gl.createBuffer();
+        colorBufferRef.current = colorBuffer;
+        gl.bindBuffer(gl.ARRAY_BUFFER, colorBuffer);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(colors), gl.STATIC_DRAW);
+    
+        const indexBuffer = gl.createBuffer();
+        indexBufferRef.current = indexBuffer;
+        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, indexBuffer);
+        gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(indices), gl.STATIC_DRAW);        
+    }
+
+    React.useEffect(() => {
+        createCube();
+    }, []);
+
+    useAnimationFrame(draw);
+};
+
+export default useCube;
